changed fuse feature structure
