+++ zerostash-fuse/Cargo.toml
@@ -6,5 +6,12 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-infinitree = { version = "0.9.0", features = ["cryptobox", "yubikey"] }
+infinitree = { version = "0.10.0", features = ["cryptobox", "yubikey"] }
+zerostash-files = { version = "0.6.0", path = "../zerostash-files" }
 serde = { version = "1.0.145" }
+scc = "1.1.0"
+fuse_mt = "0.6.0"
+tracing = "0.1.37"
+ctrlc = "3.2.3"
+nix = { version = "0.24.2", default-features = false, features = ["user"] }
+anyhow = "1.0.65"
