+++ zerostash-fuse/src/lib.rs
@@ -1,14 +1,1 @@
-#![deny(
-    arithmetic_overflow,
-    future_incompatible,
-    nonstandard_style,
-    rust_2018_idioms,
-    trivial_casts,
-    unused_crate_dependencies,
-    unused_lifetimes,
-    unused_qualifications
-)]
-
-pub mod directory;
-pub mod dir;
-pub mod files;
+pub mod mount;
