+++ zerostash/Cargo.toml
@@ -26,7 +26,6 @@ bech32 = "0.9.1"
 
 dirs = "5.0.1"
 xdg = "2.5.0"
-scc = "1.8.0"
 
 async-trait = "0.1.68"
 humansize = "2.1.3"
@@ -36,18 +35,11 @@ nix = { version = "0.26.2", default-features = false, features = ["user"] }
 abscissa_tokio= "0.7.0"
 abscissa_core= "0.7.0"
 regex = "1.8.1"
-fuse_mt = "0.6.0"
-ctrlc = "3.2.3"
-tracing = "0.1.37"
 
 secrecy = { version = "0.8.0", features = ["serde"] }
-scc = {version = "1.1.0", optional = true}
-fuse_mt = {version = "0.6.0", optional = true}
-tracing = {version = "0.1.37", optional = true}
-ctrlc = {version = "3.2.3", optional = true}
 
 [features]
-fuse = ["dep:zerostash-fuse", "dep:scc", "dep:fuse_mt", "dep:ctrlc", "dep:tracing"]
+fuse = ["dep:zerostash-fuse"]
 
 [target.'cfg(target_os = "macos")'.dependencies]
 security-framework = "2.8.2"
